Move async logic to helper file
